Conflicts:
	build.gradle
	src/main/groovy/com/entagen/jenkins/JenkinsJobManager.groovy
	src/main/groovy/com/entagen/jenkins/Main.groovy

+++ src/main/groovy/com/entagen/jenkins/JenkinsJobManager.groovy
@@ -14,6 +14,7 @@ class JenkinsJobManager {
     
     Boolean dryRun = false
     Boolean noViews = false
+    Boolean noDelete = false
 
     JenkinsApi jenkinsApi
     GitApi gitApi
@@ -48,7 +49,9 @@ class JenkinsJobManager {
         List<ConcreteJob> expectedJobs = this.expectedJobs(templateJobs, nonTemplateBranchNames)
 
         createMissingJobs(expectedJobs, currentTemplateDrivenJobNames, templateJobs)
-        deleteDeprecatedJobs(currentTemplateDrivenJobNames - expectedJobs.jobName)
+        if (!noDelete) {
+            deleteDeprecatedJobs(currentTemplateDrivenJobNames - expectedJobs.jobName)
+        }
     }
 
     public void createMissingJobs(List<ConcreteJob> expectedJobs, List<String> currentJobs, List<TemplateJob> templateJobs) {
@@ -108,8 +111,10 @@ class JenkinsJobManager {
         List<BranchView> missingBranchViews = expectedBranchViews.findAll { BranchView branchView -> !existingViewNames.contains(branchView.viewName)}
         addMissingViews(missingBranchViews)
 
-        List<String> deprecatedViewNames = getDeprecatedViewNames(existingViewNames, expectedBranchViews)
-        deleteDeprecatedViews(deprecatedViewNames)
+        if (!noDelete) {
+            List<String> deprecatedViewNames = getDeprecatedViewNames(existingViewNames, expectedBranchViews)
+            deleteDeprecatedViews(deprecatedViewNames)
+        }
     }
 
     public void addMissingViews(List<BranchView> missingViews) {
